Create upgradeable.js
In this example, we have defined a single function upgradeContract that takes the new contract address as input. Inside the function, we first retrieve the accounts using web3.eth.getAccounts(). Then, we access the deployed contract instance using Contract.deployed(). Finally, we call the upgrade function on the contract instance, passing the new contract address and the from address as options.


Note that Contract should be replaced with the actual contract object or contract class depending on your project.


By exporting the upgradeContract function, it becomes available for other files to import and use.