Create interfaces.sol
In this example, we have defined two interfaces: IERC20 and IUniswapRouter.


The IERC20 interface provides the balanceOf function, which is used to retrieve the token balance of a specific account.


The IUniswapRouter interface provides the swapExactTokensForTokens function, which is used to execute a token swap on the Uniswap or Sushiswap router. The function takes the amount of tokens to be swapped (amountIn), the minimum amount of tokens expected to be received (amountOutMin), the token path for the swap (path), the recipient address (to), and the deadline for the swap (deadline).

In this expanded version, we have added additional functions to the IERC20 interface:



transfer: Allows the contract to transfer tokens from the contract to a recipient address.

approve: Allows the contract to approve a spender to transfer tokens on behalf of the owner.

allowance: Returns the amount of tokens that the spender is allowed to transfer on behalf of the owner.


For the IUniswapRouter interface, we have added two additional functions:



getAmountsOut: Returns the expected output amounts for a given input amount and token path.

addLiquidity and removeLiquidity: Functions for adding and removing liquidity from a Uniswap pool.


These additional functions provide more flexibility and functionality for interacting with ERC20 tokens and the Uniswap router.